refactor: restructure RNIterableAPIModule to delegate functionality to RNIterableAPIModuleImpl
